refac: GlobalExceptionHandler community.handler -> common.handler 패키지로 수정
